python_service/requirements.txt
docling[all]
PyMuPDF==1.23.23
python-docx==1.1.2
mammoth==1.7.0
markdownify==0.11.6
grpcio==1.62.1
grpcio-tools==1.62.1
